Organized plugins, imports and added keymaps + configured LSP.
- To remove unwanted plugins
- To organized imports in `init.lua` files.
- Add keymaps that made more sense.

gh.nvim was installed but wasn't in use so I deleted it. Organized
imports in `init.lua` files by grouping them by category.

Configured undotre to have the window open up on the right instead of
left to not conflict with the left pad opened by no neck pain.

Added/updated keymaps for octo, defaults, and undotree.

Made better changes for LSP. Such as code to create a floating window
for renaming variables like in coc.nvim. Also downloaded nvim-navic with
integration with lualine for breadcrumbs to see if breadcrumbs is
necessary.